split c prgram
