+++ schloesser/exercise05_06_07/bonus_c/.gitignore
@@ -1,5 +1,7 @@
 *.swp
 main
 
-data/*
-data/
+data*
+
+html*
+latex*
