+++ schloesser/exercise05_06_07/bonus_c/Makefile
@@ -1,19 +1,19 @@
 CFLAGS= -std=c11 -Wall -g
 CC=gcc
-OBJ=
-HEADERS=
+OBJ=main.o graph.o
+HEADERS=graph.h
 TARGET=main
 
 default: $(TARGET)
 all: default
 
-.PHONY: clean all default run $(TARGET) valgrind
+.PHONY: clean all default run $(TARGET) valgrind doc
 
 %.o: %.c $(HEADERS)
 	$(CC) $(CFLAGS) -c $< -o $@
 
-$(TARGET):
-	$(CC) $(CFLAGS) main.c -o $(TARGET)
+$(TARGET): $(OBJ)
+	$(CC) $(CFLAGS) -o $(TARGET) $^
 
 run: $(TARGET)
 	./main ../data/small.gph
@@ -23,4 +23,7 @@ valgrind: $(TARGET)
 
 clean:
 	rm -f *.o main
+	rm -r html
 
+doc:
+	doxygen dox-conf
